Extracted State-Components out of the App + moved pipes to ui-components library
