+++ projects/state-components/README.md
@@ -1,0 +1,24 @@
+# StateComponents
+
+This library was generated with [Angular CLI](https://github.com/angular/angular-cli) version 11.2.9.
+
+## Code scaffolding
+
+Run `ng generate component component-name --project state-components` to generate a new component. You can also use `ng generate directive|pipe|service|class|guard|interface|enum|module --project state-components`.
+> Note: Don't forget to add `--project state-components` or else it will be added to the default project in your `angular.json` file. 
+
+## Build
+
+Run `ng build state-components` to build the project. The build artifacts will be stored in the `dist/` directory.
+
+## Publishing
+
+After building your library with `ng build state-components`, go to the dist folder `cd dist/state-components` and run `npm publish`.
+
+## Running unit tests
+
+Run `ng test state-components` to execute the unit tests via [Karma](https://karma-runner.github.io).
+
+## Further help
+
+To get more help on the Angular CLI use `ng help` or go check out the [Angular CLI Overview and Command Reference](https://angular.io/cli) page.
