+++ projects/state-components/ng-package.json
@@ -1,0 +1,7 @@
+{
+  "$schema": "../../node_modules/ng-packagr/ng-package.schema.json",
+  "dest": "../../dist/state-components",
+  "lib": {
+    "entryFile": "src/projects.ts"
+  }
+}
