+++ projects/state-components/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "state-components",
+  "version": "0.0.1",
+  "peerDependencies": {
+    "@angular/common": "^11.2.9",
+    "@angular/core": "^11.2.9"
+  },
+  "dependencies": {
+    "tslib": "^2.0.0"
+  }
+}
