+++ projects/state-components/src/lib/clip-chips-list/clip-chips-list.component.scss
