+++ projects/state-components/src/lib/clip-preview/clip-preview.component.ts
