+++ projects/state-components/src/lib/clip-type/clip-type.component.ts
