+++ projects/state-components/src/lib/compact-clip-card/compact-clip-card.component.ts
