+++ projects/state-components/src/projects.ts
@@ -1,0 +1,16 @@
+/*
+ * Public API Surface of state-components
+ */
+
+export * from './lib/clip-preview/clip-preview.module';
+export * from './lib/clip-preview/clip-preview.component';
+
+export * from './lib/clip-type/clip-type.module';
+export * from './lib/clip-type/clip-type.component';
+
+export * from './lib/compact-clip-card/compact-clip-card.module';
+export * from './lib/compact-clip-card/compact-clip-card.component';
+
+
+export * from './lib/clip-chips-list/clip-chips-list.module';
+export * from './lib/clip-chips-list/clip-chips-list.component';
