+++ projects/ui-components/src/lib/pipes/pipes.module.ts
