+++ projects/ui-components/src/lib/pipes/position-to-string.pipe.ts
