+++ projects/ui-components/src/lib/pipes/settings-to-sizing-type.pipe.ts
