+++ projects/ui-components/src/lib/pipes/variable-converter.pipe.spec.ts
@@ -1,0 +1,8 @@
+import {VariableConverterPipe} from './variable-converter.pipe';
+
+describe('VariableConverterPipe', () => {
+  it('create an instance', () => {
+    const pipe = new VariableConverterPipe();
+    expect(pipe).toBeTruthy();
+  });
+});
