+++ projects/ui-components/src/projects.ts
@@ -4,3 +4,4 @@
 
 export * from './lib/hint-panel/hint-panel.component';
 export * from './lib/hint-panel/hint-panel.module';
+export * from './lib/pipes/pipes.module';
